% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_footer}
\alias{create_footer}
\title{create_footer
From https://github.com/bbc/bbplot/blob/master/R/finalise_plot.R}
\usage{
create_footer(source_name, logo_image_path)
}
\arguments{
\item{source_name}{character: data source}

\item{logo_image_path}{}
}
\value{

}
\description{
create_footer
From https://github.com/bbc/bbplot/blob/master/R/finalise_plot.R
}
