% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{finalise_plot}
\alias{finalise_plot}
\title{Arrange alignment and add logo
From https://github.com/bbc/bbplot/blob/master/R/finalise_plot.R
It will left align your title, subtitle and source, add the logo at the bottom right}
\usage{
finalise_plot(
  plot_name,
  source_name = "Source = World Bank",
  width_pixels = 640,
  height_pixels = 450,
  logo_image_path = system.file("app/www", "logo.png", package =
    "pakEducationDashboard")
)
}
\arguments{
\item{plot_name}{The variable name of the plot you have created that you want to format}

\item{source_name}{The text you want to come after the text 'Source:' in the bottom left hand side of your side}

\item{width_pixels}{Width in pixels that you want to save your chart to - defaults to 640}

\item{height_pixels}{Height in pixels that you want to save your chart to - defaults to 450}

\item{logo_image_path}{File path for the logo image you want to use in the right hand side of your chart,
which needs to be a PNG file - defaults to BBC blocks image that sits within the data folder of your package}
}
\value{
(invisibly) an updated ggplot object.
}
\description{
Arrange alignment and add logo
From https://github.com/bbc/bbplot/blob/master/R/finalise_plot.R
It will left align your title, subtitle and source, add the logo at the bottom right
}
\examples{
finalise_plot(plot_name = myplot,
source = "The source for my data",
width_pixels = 640,
height_pixels = 450,
logo_image_path = "logo_image_filepath.png"
)

}
