% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{left_align}
\alias{left_align}
\title{left_align
From https://github.com/bbc/bbplot/blob/master/R/finalise_plot.R}
\usage{
left_align(plot_name, pieces)
}
\arguments{
\item{plot_name}{ggplot: Plot name}

\item{pieces}{character: Plot elements to be left aligned}
}
\value{
grob
}
\description{
left_align
From https://github.com/bbc/bbplot/blob/master/R/finalise_plot.R
}
