% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_district_map.R
\name{mod_district_map_ui}
\alias{mod_district_map_ui}
\alias{mod_district_map_server}
\title{mod_district_map_ui and mod_district_map_server}
\usage{
mod_district_map_ui(id)

mod_district_map_server(input, output, session, selection_vars)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
