% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_faq.R
\name{mod_faq_ui}
\alias{mod_faq_ui}
\alias{mod_faq_server}
\title{mod_faq_ui and mod_faq_server}
\usage{
mod_faq_ui(id)

mod_faq_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
