% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_lines}
\alias{plot_lines}
\title{plot_lines
Common ggplot layers for the line plots}
\usage{
plot_lines(
  data,
  wght_data = NULL,
  x,
  y,
  color,
  dataset,
  group,
  year,
  tooltip_value,
  font_size = 20,
  point_size = 5,
  line_size = 1.6,
  title,
  subtitle
)
}
\arguments{
\item{data}{data.frame: The dataset to be plotted}

\item{wght_data}{data.frame: The dataset containing weighted-mix data}

\item{x}{object: Variable to be mapped to x aesthetic}

\item{y}{object: Variable to be mapped to y aesthetic}

\item{color}{object: Variable to be mapped to color aesthetic}

\item{dataset}{object: Variable to be displayed in tooltip}

\item{group}{object: Variable used for dis-aggregation & to be displayed in tooltip}

\item{year}{object: Variable to be displayed in tooltip}

\item{tooltip_value}{object: Variable to be displayed in tooltip}
}
\value{

}
\description{
plot_lines
Common ggplot layers for the line plots
}
