% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_line_charts_province.R
\name{mod_line_charts_province_ui}
\alias{mod_line_charts_province_ui}
\alias{mod_line_charts_province_server}
\title{mod_line_charts_province_ui and mod_line_charts_province_server}
\usage{
mod_line_charts_province_ui(id)

mod_line_charts_province_server(input, output, session, selection_vars)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
