% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_lines}
\alias{plot_lines}
\title{plot_lines
Common ggplot layers for the line plots using non-weighted data}
\usage{
plot_lines(data, x, y, color, dataset, gender, year, tooltip_value)
}
\arguments{
\item{data}{data.frame: The dataset to be plotted}

\item{x}{object: Variable to be mapped to x aesthetic}

\item{y}{object: Variable to be mapped to y aesthetic}

\item{color}{object: Variable to be mapped to color aesthetic}

\item{dataset}{object: Variable to be displayed in tooltip}

\item{gender}{object: Variable to be displayed in tooltip}

\item{year}{object: Variable to be displayed in tooltip}

\item{tooltip_value}{object: Variable to be displayed in tooltip}
}
\value{

}
\description{
plot_lines
Common ggplot layers for the line plots using non-weighted data
}
